refactor(rag): split RAG files and tests under 300 line limit
