+++ internal/rag/automigrate_test.go
@@ -30,10 +30,6 @@ func freshMigrate(t *testing.T) *gorm.DB {
 	return testhelpers.ConnectTestDB(t)
 }
 
-// --------------------------------------------------------------------------
-// 1. TestAutoMigrate_CreatesEveryTable
-// --------------------------------------------------------------------------
-
 func TestAutoMigrate_CreatesEveryTable(t *testing.T) {
 	db := freshMigrate(t)
 
@@ -71,16 +67,6 @@ func TestAutoMigrate_CreatesEveryTable(t *testing.T) {
 			names, expected)
 	}
 }
-
-// --------------------------------------------------------------------------
-// 2. TestAutoMigrate_CreatesEveryExpectedIndex
-// --------------------------------------------------------------------------
-//
-// Uses a must-exist-set (plus shape assertions for the load-bearing
-// partial/GIN indexes). We deliberately accept "extra" indexes that gorm
-// may auto-create for FK columns — those are benign. The business-value
-// assertion is that the hand-crafted indexes the plan mandates are
-// present with the correct shape.
 
 func TestAutoMigrate_CreatesEveryExpectedIndex(t *testing.T) {
 	db := freshMigrate(t)
@@ -172,19 +158,6 @@ func TestAutoMigrate_CreatesEveryExpectedIndex(t *testing.T) {
 	checkContains("idx_rag_index_attempt_heartbeat", "'in_progress'")
 }
 
-// --------------------------------------------------------------------------
-// 3. TestAutoMigrate_AllFKConstraintsInPlace
-// --------------------------------------------------------------------------
-
-type fkRow struct {
-	ConstraintName string `gorm:"column:constraint_name"`
-	TableName      string `gorm:"column:table_name"`
-	ColumnName     string `gorm:"column:column_name"`
-	ForeignTable   string `gorm:"column:foreign_table"`
-	ForeignColumn  string `gorm:"column:foreign_column"`
-	DeleteRule     string `gorm:"column:delete_rule"`
-}
-
 func TestAutoMigrate_AllFKConstraintsInPlace(t *testing.T) {
 	db := freshMigrate(t)
 
@@ -273,10 +246,6 @@ func TestAutoMigrate_AllFKConstraintsInPlace(t *testing.T) {
 	}
 }
 
-// --------------------------------------------------------------------------
-// 4. TestAutoMigrate_Idempotent
-// --------------------------------------------------------------------------
-
 func TestAutoMigrate_Idempotent(t *testing.T) {
 	db := freshMigrate(t)
 
@@ -301,435 +270,5 @@ func TestAutoMigrate_Idempotent(t *testing.T) {
 	if before != after {
 		t.Fatalf("idempotence broken: index count changed %d → %d", before, after)
 	}
-}
-
-// --------------------------------------------------------------------------
-// 5. TestAutoMigrate_OrgFullCascadeDelete
-//
-// GDPR-critical: deleting an org must wipe every RAG row that org owns,
-// whether directly (org_id FK) or transitively (via in_connection_id FK
-// that itself cascades from org).
-// --------------------------------------------------------------------------
-
-func TestAutoMigrate_OrgFullCascadeDelete(t *testing.T) {
-	db := freshMigrate(t)
-
-	org := testhelpers.NewTestOrg(t, db)
-	user := testhelpers.NewTestUser(t, db, org.ID)
-	integ := testhelpers.NewTestInIntegration(t, db, "github")
-	conn := testhelpers.NewTestInConnection(t, db, org.ID, user.ID, integ.ID)
-	src := testhelpers.NewTestRAGSource(t, db, org.ID, conn.ID)
-
-	// RAGDocument
-	doc := &ragmodel.RAGDocument{
-		ID:           "test-doc-" + uuid.NewString(),
-		OrgID:        org.ID,
-		SemanticID:   "Test Doc",
-		LastModified: time.Now(),
-	}
-	if err := db.Create(doc).Error; err != nil {
-		t.Fatalf("create doc: %v", err)
-	}
-
-	// RAGHierarchyNode
-	node := &ragmodel.RAGHierarchyNode{
-		OrgID:       org.ID,
-		RawNodeID:   "raw-" + uuid.NewString(),
-		DisplayName: "Folder",
-		Source:      ragmodel.DocumentSourceGithub,
-		NodeType:    ragmodel.HierarchyNodeTypeFolder,
-	}
-	if err := db.Create(node).Error; err != nil {
-		t.Fatalf("create node: %v", err)
-	}
-
-	// RAGDocumentBySource
-	dbc := &ragmodel.RAGDocumentBySource{
-		DocumentID:     doc.ID,
-		RAGSourceID:    src.ID,
-		HasBeenIndexed: false,
-	}
-	if err := db.Create(dbc).Error; err != nil {
-		t.Fatalf("create doc-by-source: %v", err)
-	}
-
-	// RAGHierarchyNodeBySource
-	hbc := &ragmodel.RAGHierarchyNodeBySource{
-		HierarchyNodeID: node.ID,
-		RAGSourceID:     src.ID,
-	}
-	if err := db.Create(hbc).Error; err != nil {
-		t.Fatalf("create hier-by-source: %v", err)
-	}
-
-	// RAGIndexAttempt + error
-	attempt := &ragmodel.RAGIndexAttempt{
-		OrgID:       org.ID,
-		RAGSourceID: src.ID,
-		Status:      ragmodel.IndexingStatusNotStarted,
-	}
-	if err := db.Create(attempt).Error; err != nil {
-		t.Fatalf("create attempt: %v", err)
-	}
-	iaErr := &ragmodel.RAGIndexAttemptError{
-		OrgID:          org.ID,
-		IndexAttemptID: attempt.ID,
-		RAGSourceID:    src.ID,
-		FailureMessage: "synthetic failure for cascade test",
-	}
-	if err := db.Create(iaErr).Error; err != nil {
-		t.Fatalf("create attempt error: %v", err)
-	}
-
-	// RAGSyncRecord
-	sr := &ragmodel.RAGSyncRecord{
-		OrgID:         org.ID,
-		EntityID:      conn.ID,
-		SyncType:      ragmodel.SyncTypePruning,
-		SyncStatus:    ragmodel.SyncStatusInProgress,
-		SyncStartTime: time.Now(),
-	}
-	if err := db.Create(sr).Error; err != nil {
-		t.Fatalf("create sync record: %v", err)
-	}
-
-	// RAGSyncState
-	ss := &ragmodel.RAGSyncState{
-		OrgID:          org.ID,
-		RAGSourceID:    src.ID,
-		Status:         ragmodel.RAGConnectionStatusActive,
-		AccessType:     ragmodel.AccessTypePrivate,
-		ProcessingMode: ragmodel.ProcessingModeRegular,
-	}
-	if err := db.Create(ss).Error; err != nil {
-		t.Fatalf("create sync state: %v", err)
-	}
-
-	// RAGSearchSettings — need a real embedding model id from seed.
-	settings := &ragmodel.RAGSearchSettings{
-		OrgID:              org.ID,
-		EmbeddingModelID:   "siliconflow:qwen3-embedding-4b",
-		EmbeddingDim:       2560,
-		EmbeddingPrecision: ragmodel.EmbeddingPrecisionFloat,
-		IndexName:          "rag_chunks__siliconflow_qwen3_embedding_4b__2560",
-	}
-	if err := db.Create(settings).Error; err != nil {
-		t.Fatalf("create search settings: %v", err)
-	}
-
-	// RAGExternalUserGroup
-	eug := &ragmodel.RAGExternalUserGroup{
-		OrgID:               org.ID,
-		RAGSourceID:         src.ID,
-		ExternalUserGroupID: "github_backend-cascade-" + uuid.NewString(),
-		DisplayName:         "Backend",
-	}
-	if err := db.Create(eug).Error; err != nil {
-		t.Fatalf("create external user group: %v", err)
-	}
-
-	// RAGUserExternalUserGroup
-	uug := &ragmodel.RAGUserExternalUserGroup{
-		UserID:              user.ID,
-		ExternalUserGroupID: eug.ExternalUserGroupID,
-		RAGSourceID:         src.ID,
-	}
-	if err := db.Create(uug).Error; err != nil {
-		t.Fatalf("create user-external-group: %v", err)
-	}
-
-	// RAGPublicExternalUserGroup
-	pug := &ragmodel.RAGPublicExternalUserGroup{
-		ExternalUserGroupID: "github_public-cascade-" + uuid.NewString(),
-		RAGSourceID:         src.ID,
-	}
-	if err := db.Create(pug).Error; err != nil {
-		t.Fatalf("create public-external-group: %v", err)
-	}
-
-	// RAGExternalIdentity
-	ident := &ragmodel.RAGExternalIdentity{
-		OrgID:              org.ID,
-		UserID:             user.ID,
-		RAGSourceID:        src.ID,
-		Provider:           "github",
-		ExternalUserID:     "gh-" + uuid.NewString(),
-		ExternalUserEmails: pq.StringArray{"alice@example.com"},
-	}
-	if err := db.Create(ident).Error; err != nil {
-		t.Fatalf("create external identity: %v", err)
-	}
-
-	// Snapshot the src.ID before deleting org — after cascade,
-	// rag_sources cascades off org, which cascades into every RAG
-	// junction table that keys off rag_source_id.
-	srcID := src.ID
-
-	// Hiveloop's own orgs → org_memberships FK does NOT cascade
-	// (owned by internal/model, not our concern). Remove memberships
-	// first; RAG's own org cascade is what this test is exercising.
-	if err := db.Exec("DELETE FROM org_memberships WHERE org_id = ?", org.ID).Error; err != nil {
-		t.Fatalf("delete org memberships: %v", err)
-	}
-
-	// The bomb: delete the org.
-	if err := db.Exec("DELETE FROM orgs WHERE id = ?", org.ID).Error; err != nil {
-		t.Fatalf("delete org: %v", err)
-	}
-
-	// Now verify every RAG row is gone. Query by org_id where it exists,
-	// by rag_source_id where it doesn't.
-	checks := []struct {
-		table string
-		where string
-		args  []any
-	}{
-		{"rag_documents", "org_id = ?", []any{org.ID}},
-		{"rag_hierarchy_nodes", "org_id = ?", []any{org.ID}},
-		{"rag_document_by_sources", "rag_source_id = ?", []any{srcID}},
-		{"rag_hierarchy_node_by_sources", "rag_source_id = ?", []any{srcID}},
-		{"rag_index_attempts", "org_id = ?", []any{org.ID}},
-		{"rag_index_attempt_errors", "org_id = ?", []any{org.ID}},
-		{"rag_sync_records", "org_id = ?", []any{org.ID}},
-		{"rag_sync_states", "org_id = ?", []any{org.ID}},
-		{"rag_sources", "org_id = ?", []any{org.ID}},
-		{"rag_search_settings", "org_id = ?", []any{org.ID}},
-		{"rag_external_user_groups", "org_id = ?", []any{org.ID}},
-		{"rag_user_external_user_groups", "rag_source_id = ?", []any{srcID}},
-		{"rag_public_external_user_groups", "rag_source_id = ?", []any{srcID}},
-		{"rag_external_identities", "org_id = ?", []any{org.ID}},
-	}
-	for _, c := range checks {
-		var n int64
-		sql := "SELECT COUNT(*) FROM " + c.table + " WHERE " + c.where
-		if err := db.Raw(sql, c.args...).Scan(&n).Error; err != nil {
-			t.Fatalf("post-delete count on %s: %v", c.table, err)
-		}
-		if n != 0 {
-			t.Errorf("%s: %d row(s) survived org cascade delete", c.table, n)
-		}
-	}
-}
-
-// --------------------------------------------------------------------------
-// 6. TestAutoMigrate_SeedsEmbeddingRegistry
-// --------------------------------------------------------------------------
-
-func TestAutoMigrate_SeedsEmbeddingRegistry(t *testing.T) {
-	db := freshMigrate(t)
-
-	type row struct {
-		ID        string
-		Provider  string
-		Dimension int
-	}
-	var got []row
-	if err := db.Raw(`
-		SELECT id, provider, dimension FROM rag_embedding_models
-		ORDER BY id
-	`).Scan(&got).Error; err != nil {
-		t.Fatalf("query rag_embedding_models: %v", err)
-	}
-
-	expected := []row{
-		{"openai:text-embedding-3-large", "openai", 3072},
-		{"openai:text-embedding-3-small", "openai", 1536},
-		{"siliconflow:qwen3-embedding-0.6b", "siliconflow", 1024},
-		{"siliconflow:qwen3-embedding-4b", "siliconflow", 2560},
-		{"siliconflow:qwen3-embedding-8b", "siliconflow", 4096},
-	}
-
-	if len(got) != len(expected) {
-		t.Fatalf("seed row count: got %d, want %d\n  got=%v", len(got), len(expected), got)
-	}
-	for i, want := range expected {
-		if got[i] != want {
-			t.Errorf("row[%d]: got %+v, want %+v", i, got[i], want)
-		}
-	}
-}
-
-// --------------------------------------------------------------------------
-// 7. TestAutoMigrate_PartialIndexActuallyUsed_Document
-// --------------------------------------------------------------------------
-
-func TestAutoMigrate_PartialIndexActuallyUsed_Document(t *testing.T) {
-	db := freshMigrate(t)
-
-	org := testhelpers.NewTestOrg(t, db)
-
-	now := time.Now()
-	past := now.Add(-1 * time.Hour)
-
-	// Insert 100 docs with mixed sync states — only ~1/3 match the
-	// partial-index predicate, which matches realistic production skew.
-	for i := 0; i < 100; i++ {
-		d := &ragmodel.RAGDocument{
-			ID:           "partial-doc-" + uuid.NewString(),
-			OrgID:        org.ID,
-			SemanticID:   "Doc",
-			LastModified: now,
-		}
-		switch i % 3 {
-		case 0:
-			// last_synced = NULL → matches predicate
-			d.LastSynced = nil
-		case 1:
-			// last_synced in the past → last_modified > last_synced → matches
-			d.LastSynced = &past
-		case 2:
-			// last_synced = now → last_modified == last_synced → misses
-			d.LastSynced = &now
-		}
-		if err := db.Create(d).Error; err != nil {
-			t.Fatalf("create doc %d: %v", i, err)
-		}
-	}
-	if err := db.Exec("ANALYZE rag_documents").Error; err != nil {
-		t.Fatalf("ANALYZE: %v", err)
-	}
-
-	plan := explainNoSeqScan(t, db, `
-		EXPLAIN (FORMAT JSON)
-		SELECT id FROM rag_documents
-		WHERE last_modified > last_synced OR last_synced IS NULL
-	`)
-	if !strings.Contains(plan, "idx_rag_document_needs_sync") {
-		t.Fatalf("planner did not use idx_rag_document_needs_sync; plan=%s", plan)
-	}
-}
-
-// --------------------------------------------------------------------------
-// 8. TestAutoMigrate_PartialIndexActuallyUsed_Watchdog
-// --------------------------------------------------------------------------
-
-func TestAutoMigrate_PartialIndexActuallyUsed_Watchdog(t *testing.T) {
-	db := freshMigrate(t)
-
-	org := testhelpers.NewTestOrg(t, db)
-	user := testhelpers.NewTestUser(t, db, org.ID)
-	integ := testhelpers.NewTestInIntegration(t, db, "github")
-	conn := testhelpers.NewTestInConnection(t, db, org.ID, user.ID, integ.ID)
-	src := testhelpers.NewTestRAGSource(t, db, org.ID, conn.ID)
-
-	stale := time.Now().Add(-1 * time.Hour)
-	fresh := time.Now()
-
-	// 1000 rows with ~5% in-progress. At that skew the partial
-	// heartbeat index (size ~= 50 entries) is noticeably smaller than
-	// the full-table status index (size = 1000 entries), so the planner
-	// reliably picks it once enable_seqscan is off.
-	for i := 0; i < 1000; i++ {
-		a := &ragmodel.RAGIndexAttempt{
-			OrgID:       org.ID,
-			RAGSourceID: src.ID,
-			Status:      ragmodel.IndexingStatusSuccess, // majority branch, misses partial
-		}
-		if i%20 == 0 {
-			a.Status = ragmodel.IndexingStatusInProgress
-			if i%40 == 0 {
-				a.LastProgressTime = &stale
-			} else {
-				a.LastProgressTime = &fresh
-			}
-		}
-		if err := db.Create(a).Error; err != nil {
-			t.Fatalf("create attempt %d: %v", i, err)
-		}
-	}
-	if err := db.Exec("ANALYZE rag_index_attempts").Error; err != nil {
-		t.Fatalf("ANALYZE: %v", err)
-	}
-
-	plan := explainNoSeqScan(t, db, `
-		EXPLAIN (FORMAT JSON)
-		SELECT id FROM rag_index_attempts
-		WHERE status = 'in_progress'
-		  AND last_progress_time < NOW() - INTERVAL '30 minutes'
-	`)
-	if !strings.Contains(plan, "idx_rag_index_attempt_heartbeat") {
-		t.Fatalf("planner did not use idx_rag_index_attempt_heartbeat; plan=%s", plan)
-	}
-}
-
-// --------------------------------------------------------------------------
-// 9. TestAutoMigrate_GINIndexActuallyUsed
-// --------------------------------------------------------------------------
-
-func TestAutoMigrate_GINIndexActuallyUsed(t *testing.T) {
-	db := freshMigrate(t)
-
-	org := testhelpers.NewTestOrg(t, db)
-
-	for i := 0; i < 100; i++ {
-		emails := pq.StringArray{}
-		if i%5 == 0 {
-			emails = pq.StringArray{"alice@example.com"}
-		} else {
-			emails = pq.StringArray{"bob-" + uuid.NewString() + "@example.com"}
-		}
-		d := &ragmodel.RAGDocument{
-			ID:                 "gin-doc-" + uuid.NewString(),
-			OrgID:              org.ID,
-			SemanticID:         "Doc",
-			LastModified:       time.Now(),
-			ExternalUserEmails: emails,
-		}
-		if err := db.Create(d).Error; err != nil {
-			t.Fatalf("create doc %d: %v", i, err)
-		}
-	}
-	if err := db.Exec("ANALYZE rag_documents").Error; err != nil {
-		t.Fatalf("ANALYZE: %v", err)
-	}
-
-	// The plan agent noted @> is the GIN-preferred operator; ANY(...)
-	// doesn't use GIN. We use the production shape here.
-	plan := explainNoSeqScan(t, db, `
-		EXPLAIN (FORMAT JSON)
-		SELECT id FROM rag_documents
-		WHERE external_user_emails @> ARRAY['alice@example.com']::text[]
-	`)
-	if !strings.Contains(plan, "idx_rag_document_ext_emails") {
-		t.Fatalf("planner did not use idx_rag_document_ext_emails; plan=%s", plan)
-	}
-}
-
-// --------------------------------------------------------------------------
-// Helpers
-// --------------------------------------------------------------------------
-
-// explainNoSeqScan runs the given EXPLAIN-wrapped query inside a tx with
-// enable_seqscan disabled — with <1000 rows the planner may still prefer
-// seq scan even when the index is valid. Disabling seq scan forces the
-// planner to use the best available index, which is the correct business
-// assertion: "when the query is forced through the indexed path, the
-// right index is chosen". If no usable index exists, the plan still
-// reports Seq Scan (enable_seqscan is a soft discouragement, not a
-// hard ban).
-func explainNoSeqScan(t *testing.T, db *gorm.DB, sql string) string {
-	t.Helper()
-
-	var planJSON string
-	err := db.Transaction(func(tx *gorm.DB) error {
-		if err := tx.Exec("SET LOCAL enable_seqscan = off").Error; err != nil {
-			return err
-		}
-		var plans []struct {
-			QueryPlan string `gorm:"column:QUERY PLAN"`
-		}
-		if err := tx.Raw(sql).Scan(&plans).Error; err != nil {
-			return err
-		}
-		if len(plans) == 0 {
-			return nil
-		}
-		planJSON = plans[0].QueryPlan
-		return nil
-	})
-	if err != nil {
-		t.Fatalf("explain: %v", err)
-	}
-	return planJSON
 }
