+++ internal/rag/model/document_test.go
@@ -25,10 +25,6 @@ func bootstrapDocs(t *testing.T) *gorm.DB {
 	return testhelpers.ConnectTestDB(t)
 }
 
-// cleanupDocsForOrg scopes cleanup to the org so parallel test runs
-// don't clobber each other. RAGDocument rows are the fan-out point —
-// everything below them goes away via CASCADE (junction edges) or
-// explicit deletes registered by fixture constructors.
 func cleanupDocsForOrg(t *testing.T, db *gorm.DB, orgID uuid.UUID) {
 	t.Helper()
 	t.Cleanup(func() {
@@ -45,12 +41,6 @@ func docID(t *testing.T) string {
 	t.Helper()
 	return "doc-" + uuid.NewString()
 }
-
-// ---------------------------------------------------------------------
-// 1. RAGDocument_OrgCascadeDelete
-// ---------------------------------------------------------------------
-// Business value: org deletion must tear down all its RAG rows or we
-// fail GDPR tenant-deletion compliance.
 
 func TestRAGDocument_OrgCascadeDelete(t *testing.T) {
 	db := bootstrapDocs(t)
@@ -77,13 +67,6 @@ func TestRAGDocument_OrgCascadeDelete(t *testing.T) {
 		t.Fatalf("expected ErrRecordNotFound after org cascade, got err=%v (doc=%+v)", err, found)
 	}
 }
-
-// ---------------------------------------------------------------------
-// 2. RAGDocument_ParentHierarchyNodeSetNullOnDelete
-// ---------------------------------------------------------------------
-// Business value: deleting a folder must not delete its documents —
-// the docs just lose their hierarchy parent pointer and wait for
-// re-parenting on the next sync.
 
 func TestRAGDocument_ParentHierarchyNodeSetNullOnDelete(t *testing.T) {
 	db := bootstrapDocs(t)
@@ -124,13 +107,6 @@ func TestRAGDocument_ParentHierarchyNodeSetNullOnDelete(t *testing.T) {
 		t.Fatalf("expected parent_hierarchy_node_id to be NULL after folder delete, got %d", *got.ParentHierarchyNodeID)
 	}
 }
-
-// ---------------------------------------------------------------------
-// 3. RAGDocument_NeedsSyncPartialIndexExistsAndIsUsed
-// ---------------------------------------------------------------------
-// Business value: sync loop + watchdog scan rag_documents at prod
-// scale. Without this partial index we seq-scan the whole table every
-// 15 seconds per-tenant, which is a P0.
 
 func TestRAGDocument_NeedsSyncPartialIndexExistsAndIsUsed(t *testing.T) {
 	db := bootstrapDocs(t)
@@ -198,13 +174,6 @@ func TestRAGDocument_NeedsSyncPartialIndexExistsAndIsUsed(t *testing.T) {
 	}
 }
 
-// ---------------------------------------------------------------------
-// 4. RAGDocument_GINIndexOnExternalUserEmails
-// ---------------------------------------------------------------------
-// Business value: ACL-filtered retrieval. We mirror email membership
-// filters in LanceDB but also run them in Postgres (perm-sync, audit,
-// admin UI). Full-scan ACL filter on a 1M-doc tenant = timeouts.
-
 func TestRAGDocument_GINIndexOnExternalUserEmails(t *testing.T) {
 	db := bootstrapDocs(t)
 	org := testhelpers.NewTestOrg(t, db)
@@ -261,240 +230,3 @@ func TestRAGDocument_GINIndexOnExternalUserEmails(t *testing.T) {
 	}
 }
 
-// ---------------------------------------------------------------------
-// 5. RAGHierarchyNode_UniqueRawIDSource
-// ---------------------------------------------------------------------
-// Business value: prevent the same upstream page being ingested under
-// two different HierarchyNode rows (cache-invalidation hell; user
-// sees duplicates in tree).
-
-func TestRAGHierarchyNode_UniqueRawIDSource(t *testing.T) {
-	db := bootstrapDocs(t)
-	org := testhelpers.NewTestOrg(t, db)
-	cleanupDocsForOrg(t, db, org.ID)
-
-	raw := "collision-raw-" + uuid.NewString()
-	first := &ragmodel.RAGHierarchyNode{
-		OrgID:       org.ID,
-		RawNodeID:   raw,
-		DisplayName: "A",
-		Source:      ragmodel.DocumentSourceConfluence,
-		NodeType:    ragmodel.HierarchyNodeTypeSpace,
-	}
-	if err := db.Create(first).Error; err != nil {
-		t.Fatalf("first insert: %v", err)
-	}
-
-	second := &ragmodel.RAGHierarchyNode{
-		OrgID:       org.ID,
-		RawNodeID:   raw,
-		DisplayName: "B",
-		Source:      ragmodel.DocumentSourceConfluence,
-		NodeType:    ragmodel.HierarchyNodeTypeSpace,
-	}
-	err := db.Create(second).Error
-	if err == nil {
-		t.Fatalf("expected unique-violation on duplicate (raw_node_id, source); got nil")
-	}
-	if !strings.Contains(err.Error(), "uq_rag_hierarchy_node_raw_id_source") {
-		t.Fatalf("error did not mention the unique index: %v", err)
-	}
-}
-
-// ---------------------------------------------------------------------
-// 6. RAGDocumentBySource_SourceCascade
-// ---------------------------------------------------------------------
-// Business value: removing a RAGSource must sweep its junction edges
-// but must NOT remove the underlying document (which may still be
-// referenced by a second source).
-
-func TestRAGDocumentBySource_SourceCascade(t *testing.T) {
-	db := bootstrapDocs(t)
-	org := testhelpers.NewTestOrg(t, db)
-	cleanupDocsForOrg(t, db, org.ID)
-
-	user := testhelpers.NewTestUser(t, db, org.ID)
-	integ := testhelpers.NewTestInIntegration(t, db, "github")
-	conn := testhelpers.NewTestInConnection(t, db, org.ID, user.ID, integ.ID)
-	src := testhelpers.NewTestRAGSource(t, db, org.ID, conn.ID)
-
-	doc := &ragmodel.RAGDocument{
-		ID:           docID(t),
-		OrgID:        org.ID,
-		SemanticID:   "Shared Doc",
-		LastModified: time.Now(),
-	}
-	if err := db.Create(doc).Error; err != nil {
-		t.Fatalf("create doc: %v", err)
-	}
-	edge := &ragmodel.RAGDocumentBySource{
-		DocumentID:     doc.ID,
-		RAGSourceID:    src.ID,
-		HasBeenIndexed: true,
-	}
-	if err := db.Create(edge).Error; err != nil {
-		t.Fatalf("create edge: %v", err)
-	}
-
-	if err := db.Exec(`DELETE FROM rag_sources WHERE id = ?`, src.ID).Error; err != nil {
-		t.Fatalf("delete rag_source: %v", err)
-	}
-
-	// Edge gone.
-	var edgeCount int64
-	if err := db.Model(&ragmodel.RAGDocumentBySource{}).
-		Where("document_id = ? AND rag_source_id = ?", doc.ID, src.ID).
-		Count(&edgeCount).Error; err != nil {
-		t.Fatalf("count edges: %v", err)
-	}
-	if edgeCount != 0 {
-		t.Fatalf("expected 0 edge rows after rag_source delete, got %d", edgeCount)
-	}
-
-	// Doc survives.
-	var docCount int64
-	if err := db.Model(&ragmodel.RAGDocument{}).
-		Where("id = ?", doc.ID).
-		Count(&docCount).Error; err != nil {
-		t.Fatalf("count docs: %v", err)
-	}
-	if docCount != 1 {
-		t.Fatalf("doc should survive rag_source delete, got count=%d", docCount)
-	}
-}
-
-// ---------------------------------------------------------------------
-// 7. HierarchyNodeType_IsValid (pure)
-// ---------------------------------------------------------------------
-// Business value: API input validation — admin UI must reject a
-// client-supplied node_type of "foo" before it corrupts the ingest
-// pipeline.
-
-func TestRAGHierarchyNodeType_IsValid(t *testing.T) {
-	cases := []struct {
-		in   ragmodel.HierarchyNodeType
-		want bool
-	}{
-		{ragmodel.HierarchyNodeTypeFolder, true},
-		{ragmodel.HierarchyNodeTypeSource, true},
-		{ragmodel.HierarchyNodeTypeSharedDrive, true},
-		{ragmodel.HierarchyNodeTypeMyDrive, true},
-		{ragmodel.HierarchyNodeTypeSpace, true},
-		{ragmodel.HierarchyNodeTypePage, true},
-		{ragmodel.HierarchyNodeTypeProject, true},
-		{ragmodel.HierarchyNodeTypeDatabase, true},
-		{ragmodel.HierarchyNodeTypeWorkspace, true},
-		{ragmodel.HierarchyNodeTypeSite, true},
-		{ragmodel.HierarchyNodeTypeDrive, true},
-		{ragmodel.HierarchyNodeTypeChannel, true},
-		{ragmodel.HierarchyNodeType(""), false},
-		{ragmodel.HierarchyNodeType("random_string"), false},
-		{ragmodel.HierarchyNodeType("Folder"), false}, // case-sensitive per Onyx
-		{ragmodel.HierarchyNodeType(" folder"), false},
-	}
-	for _, c := range cases {
-		got := c.in.IsValid()
-		if got != c.want {
-			t.Errorf("HierarchyNodeType(%q).IsValid() = %v, want %v", c.in, got, c.want)
-		}
-	}
-}
-
-// ---------------------------------------------------------------------
-// 8. DocumentSource_IsValid (pure)
-// ---------------------------------------------------------------------
-func TestDocumentSource_IsValid(t *testing.T) {
-	// Spot-check across the enum: first, last-from-Onyx-range, special
-	// cases, and a couple of mid-range ones.
-	positives := []ragmodel.DocumentSource{
-		ragmodel.DocumentSourceIngestionAPI,
-		ragmodel.DocumentSourceSlack,
-		ragmodel.DocumentSourceGithub,
-		ragmodel.DocumentSourceConfluence,
-		ragmodel.DocumentSourceGoogleDrive,
-		ragmodel.DocumentSourceNotion,
-		ragmodel.DocumentSourceS3,
-		ragmodel.DocumentSourceGoogleCloudStorage,
-		ragmodel.DocumentSourceNotApplicable,
-		ragmodel.DocumentSourceMockConnector,
-		ragmodel.DocumentSourceCraftFile,
-		ragmodel.DocumentSourceBitbucket,
-		ragmodel.DocumentSourceTestrail,
-	}
-	for _, p := range positives {
-		if !p.IsValid() {
-			t.Errorf("expected %q to be valid", p)
-		}
-	}
-	negatives := []ragmodel.DocumentSource{
-		"",
-		"random_source",
-		"Slack",
-		"google-drive",
-		" slack",
-		"jira ",
-	}
-	for _, n := range negatives {
-		if n.IsValid() {
-			t.Errorf("expected %q to be invalid", n)
-		}
-	}
-}
-
-// ---------------------------------------------------------------------
-// helpers: EXPLAIN plan inspection
-// ---------------------------------------------------------------------
-
-// explainJSON returns the raw text of EXPLAIN (FORMAT JSON) for q.
-// Rather than parse the JSON tree we do substring matches against
-// the known index name — robust to Postgres version changes in plan
-// shape but still catches a planner that picks the wrong path.
-//
-// Forces enable_seqscan=off so the planner will demonstrably pick the
-// partial/GIN index whenever one exists and covers the query — the
-// test's assertion is "there IS a usable index," not "production will
-// pick it at cost X." At the row counts a unit test can realistically
-// insert (hundreds), Postgres's default costing falls back to
-// seq-scan even when a perfectly good partial index exists. We run
-// the setting inside a transaction so it does not leak between tests.
-func explainJSON(t *testing.T, db *gorm.DB, q string, args ...any) string {
-	t.Helper()
-	var plan string
-	err := db.Transaction(func(tx *gorm.DB) error {
-		if err := tx.Exec(`SET LOCAL enable_seqscan = off`).Error; err != nil {
-			return fmt.Errorf("disable seqscan: %w", err)
-		}
-		rows, err := tx.Raw(`EXPLAIN (FORMAT JSON) `+q, args...).Rows()
-		if err != nil {
-			return fmt.Errorf("EXPLAIN: %w", err)
-		}
-		defer func() { _ = rows.Close() }()
-
-		var builder strings.Builder
-		for rows.Next() {
-			var chunk []byte
-			if err := rows.Scan(&chunk); err != nil {
-				return fmt.Errorf("scan plan: %w", err)
-			}
-			builder.Write(chunk)
-		}
-		if err := rows.Err(); err != nil {
-			return fmt.Errorf("rows.Err: %w", err)
-		}
-		plan = builder.String()
-		return nil
-	})
-	if err != nil {
-		t.Fatalf("explain tx: %v", err)
-	}
-	// Sanity-check the result is valid JSON.
-	var parsed any
-	if jerr := json.Unmarshal([]byte(plan), &parsed); jerr != nil {
-		t.Logf("warn: plan was not valid JSON: %v (text=%q)", jerr, plan)
-	}
-	return plan
-}
-
-func planMentions(plan, needle string) bool {
-	return strings.Contains(plan, needle)
-}
